# hstore: Make binary parsing 2X faster